backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.9
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
